% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMLBtravel.R
\name{getMLBtravel}
\alias{getMLBtravel}
\title{getMLBtravel}
\usage{
getMLBtravel(example)
}
\arguments{
\item{example}{SET TRUE to do example day where teams traveled rather than today's games}
}
\description{
getMLBtravel
}
