% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMLBtraveldays.R
\name{getMLBtraveldays}
\alias{getMLBtraveldays}
\title{getMLBtraveldays}
\usage{
getMLBtraveldays(day1, day2)
}
\arguments{
\item{day1}{Day to use games as origin travel locations}

\item{day2}{Day to use games as destination travel locations}
}
\description{
getMLBtraveldays
}
